fix: type check and errors
