+++ dashboard/src/common/stats-charts/components/index.ts
@@ -1,2 +1,1 @@
 export * from "./select-date-view";
-export * from "./date-x-axies";
