+++ dashboard/src/common/stats-charts/hooks/index.ts
@@ -1,0 +1,1 @@
+export * from "./use-from-now-interval"
