+++ dashboard/src/common/stats-charts/hooks/use-from-now-interval.ts
