+++ dashboard/src/common/stats-charts/index.ts
@@ -1,3 +1,4 @@
-export * from "./hooks"
 export * from "./components"
+export * from "./hooks"
+export * from "./utils"
 export * from "./types"
