+++ dashboard/src/features/total-traffic-widget/hooks/index.ts
@@ -1,2 +1,1 @@
 export * from "./use-transform-data"
-export * from "./use-from-now-interval"
